backend/requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
werkzeug
pandas
python-dotenv==1.0.1
